Move wallet specific components to the wallet directory
