+++ packages/wallet/detail/holdings/balanceInfo/unlocking.js
@@ -8,7 +8,7 @@ import {
   calculateBalanceLockedInUnvotes,
   calculateBalanceLockedInVotes,
   getActiveTokenAccount,
-} from '@wallet/utilities/account';
+} from '@wallet/utils/account';
 import Icon from '@basics/icon';
 import styles from './balanceInfo.css';
