+++ packages/wallet/detail/identity/searchBarWallets/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import WalletVisual from '@wallet/detail/identity/walletVisual';
-import { truncateAddress } from '@wallet/utilities/account';
+import WalletVisual from '@wallet/components/walletVisual';
+import { truncateAddress } from '@wallet/utils/account';
 import styles from './walletsAndDeletegates.css';
 
 const Wallets = ({
