--- packages/wallet/detail/identity/walletVisual/transactionAddress/transactionAddress.css
@@ -1,25 +1,0 @@
-@import '../../../../../../setup/react/app/mixins.css';
-
-.wrapper {
-  @mixin contentLargest bold;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  box-sizing: border-box;
-  margin: 0;
-  overflow: hidden;
-
-  & > span {
-    text-overflow: ellipsis;
-    overflow: hidden;
-    width: 100%;
-  }
-
-  & .subTitle {
-    @mixin contentNormal;
-
-    color: var(--color-slate-gray);
-  }
-}
