+++ src/modules/wallet/components/walletVisual/gradients.js
