+++ src/modules/wallet/components/walletVisual/index.js
