+++ src/modules/wallet/components/walletVisual/index.test.js
