+++ src/modules/wallet/components/walletVisual/walletVisual.css
