+++ src/modules/wallet/components/walletVisualWithAddress/walletVisualWithAddress.test.js
